Fuzz vs. std::set w/strippable code in roaring.hh
Adds code to the Roaring C++ class under #ifdef ROARING_DOUBLECHECK_CPP
which maintains a parallel state to the roaring bitmap in a `std::set`.
This code is stripped out via amalgamation.sh for a release roaring.hh

Having the code in the main class means that any codebase that uses the
C++ object can leverage the tests, so this isn't exercising a distinct
object which was created solely for testing.  It also increases the
likelihood of attention to the checks during code maintenance.

Since std::set stays sorted, the algorithms are not terribly difficult
to have written using iterators.  Though it may not be worth it vs.
using higher level facilities like `set_intersection()` that would
create temporary objects.

The corresponding std::set algorithms are implemented for the inplace
forms of the API only.  The non-inplace operators are then validated
against the inplace forms, to avoid writing the C++ code twice.

Included is a starting concept for a fuzzing unit test that uses the
ROARING_DOUBLECHECK_CPP setting.  This is how #247 was found.

The topic of fuzz testing has previously been mentioned in #122.