Move roaring_bitmap_t to roaring_types.h
The working suggestion is to be able to pass custom allocators the
handle of the roaring_bitmap for which the allocation is being
performed.  This requires it being accessible by the files that do
the allocations, like bitset.c, which do not include the full API
definition.  So move it to roaring_types.h

Includes a helper `r_from_ra()` to get a roaring_bitmap_t* from a
roaring_array_t*.