Bring C++ namespacing up to date
The bitset_t was once an internal implementation detail, but was made
a part of the api.  The roaring::internal namespace was still being
used on bitset_t.

This moves it to the API namespace.  Also, the `misc` namespace where
the configreport put its routines was not being included in some
places.  This fixes that.