ROARING_BUILD_C_TESTS_AS_CPP in add_c_benchmark()
The CMake function add_c_benchmark() did not support the flag for
building the entirety of Roaring and its tests as C++.

This adds that, and the necessary namespace inclusions (which could
be abstracted into a header file that the benchmarks include).

Also adjusts some unit tests that aren't working, as the flag for
ROARING_BUILD_C_TESTS_AS_CPP has apparently ceased to be run in
continuous integration, allowing things like this to atrophy.