Include cmocka for test before roaring, not after
The pattern in the tests was to #include whatever files were being
tested from Roaring, and then #include "test.h", which includes
the cmocka header.

This ordering gives cmocka the last vote on what to name things.  And
cmocka doesn't use any kind of naming convention or override for its
features such as calling the way to fail a test "fail()".  This means
if your codebase happens to have a use for the term fail, you have
to name it out of the way...vs. tell cmocka to call its term
`cmocka_fail()` (for example).

In order to give Roaring the final say, this pushes the inclusion of
test.h higher in the files that use it.  It also redefines the cmocka
fail()/fail_msg()/skip()->cmocka_fail()/cmocka_fail_msg()/cmocka_skip()